Moving DeploymentGenerator to its own module.
