+++ tests/cibyl/unit/plugins/openstack/sources/zuul/test_actions.py
@@ -16,89 +16,7 @@
 from unittest import TestCase
 from unittest.mock import Mock, call
 
-from cibyl.plugins.openstack.sources.zuul.actions import (DeploymentGenerator,
-                                                          DeploymentQuery)
-
-
-class TestDeploymentGenerator(TestCase):
-    """Tests for :class:`DeploymentGenerator`.
-    """
-
-    def test_no_release(self):
-        """Checks that the release is not fetched if no arg for it is passed.
-        """
-        kwargs = {
-        }
-
-        release_search = Mock()
-        release_search.search = Mock()
-
-        variant = Mock()
-
-        tools = Mock()
-        tools.release_search = release_search
-
-        generator = DeploymentGenerator(tools)
-
-        deployment = generator.generate_deployment_for(variant, **kwargs)
-
-        self.assertEqual('', deployment.release.value)
-
-        release_search.search.assert_not_called()
-
-    def test_adds_release_on_spec_arg(self):
-        """Checks that the release is fetched if the 'spec' arg is passed.
-        """
-        kwargs = {
-            'spec': None
-        }
-
-        variable = 'var'
-        release = 'v1.0'
-
-        release_search = Mock()
-        release_search.search = Mock()
-        release_search.search.return_value = (variable, release)
-
-        variant = Mock()
-
-        tools = Mock()
-        tools.release_search = release_search
-
-        generator = DeploymentGenerator(tools)
-
-        deployment = generator.generate_deployment_for(variant, **kwargs)
-
-        self.assertEqual(release, deployment.release.value)
-
-        release_search.search.assert_called_once_with(variant)
-
-    def test_adds_release_on_release_arg(self):
-        """Checks that the release is fetched if the 'release' arg is passed.
-        """
-        kwargs = {
-            'release': None
-        }
-
-        variable = 'var'
-        release = 'v1.0'
-
-        release_search = Mock()
-        release_search.search = Mock()
-        release_search.search.return_value = (variable, release)
-
-        variant = Mock()
-
-        tools = Mock()
-        tools.release_search = release_search
-
-        generator = DeploymentGenerator(tools)
-
-        deployment = generator.generate_deployment_for(variant, **kwargs)
-
-        self.assertEqual(release, deployment.release.value)
-
-        release_search.search.assert_called_once_with(variant)
+from cibyl.plugins.openstack.sources.zuul.actions import DeploymentQuery
 
 
 class TestDeploymentQuery(TestCase):
